requirements.txt
matplotlib==3.6.1
numpy==1.23.3
pandas==1.5.1
Pillow==9.5.0
plotly==5.11.0
scikit_learn==1.2.2
seaborn==0.12.2
streamlit==1.14.0
